+++ src/CommonInterface/CommonInterface.h
@@ -21,7 +21,7 @@
 
 // Note: If include of actiongenerator still exists here remove it
 // it was added for temp reasons
-#include "../Entangler/ActionGenerator.hpp"
+#include "../Walker/ActionGenerator.hpp"
 #include "../PDDLTypes/PDDLActionInstance.hpp"
 
 using namespace std;
