--- src/Entangler/Entangler.hpp
@@ -1,11 +1,0 @@
-#ifndef ENTANGLER
-#define ENTANGLER
-
-class Entangler {
-public:
-
-private:
-
-};
-
-#endif
