+++ src/Walker/DepthFunctions/ConstantDepthFunction.cpp
