+++ src/Walker/DepthFunctions/ObjectActionDepthFunction.h
@@ -1,7 +1,7 @@
 #ifndef ObjectActionDepthFunction_H
 #define ObjectActionDepthFunction_H
 
-#include "../PDDLTypes/PDDLInstance.hpp"
+#include "../../PDDLTypes/PDDLInstance.hpp"
 #include "BaseDepthFunction.h"
 
 using namespace std;
