+++ src/Walker/Heuristics/BaseHeuristics.hh
