+++ src/Walker/Heuristics/Contexts/BaseContext.h
