+++ src/Walker/Walker.hpp
@@ -1,0 +1,11 @@
+#ifndef WALKER
+#define WALKER
+
+class Walker {
+public:
+    Walker();
+private:
+
+};
+
+#endif
